devops/requirements.txt
dnspython==1.12.0
IPy==0.83
psutil==3.1.1
